sample/api/common/serializers/industry.py
from rest_framework import serializers

from sample.common.models import Industry


class IndustrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Industry
        fields = ('id', 'name')
